feat(annotations)!: new AST and full type inference
Make a minimal `AST` `ExprF` type, wrapping it with a `Cofree`-style
recursive *annotating* `struct`, which inserts arbitrary data (generic)
at each *recursion* level.
That is used to have a few different "versions" of the same `AST`,
depending on what phase we're in:
- pure parsing (`OffsetExpr`): just keeping context-free offset from the
  start of the input `&str` (two numbers)
- span attachment (`SpannedExpr`): convert the offsets to real `Noir`
  `Span`s, referring to the real place in code (both file and location
  inside of it)
- type checking (`Spanned{Optionally,}TypedExpr`): type-infer and add
  `Noir` `Type` information to all expressions (using an intermediary
  `SpannedOptionallyTypedExpr` for when we're still not sure of integer
  literals' types, as they have to come from above)

All of the conversions between the different stages are implemented as
variants of a [catamorphism](https://en.wikipedia.org/wiki/Catamorphism)
and its derivatives.

Also split all the code in separate modules (`ast`, `parse`, `typing`),
with some of the main driver logic in the top-level `lib.rs`

TODO:
- Connection with the rest of the codebase
- Quantifiers (parsing and type inference)
- Array and tuple indexing (parsing and type inference)
- Better errors (still...)
- More organization of the API itself (cleaner `lib.rs`, etc.)
- More tests