feat(fv): Update Noir dependency version
We are now getting the Noir dependency from the branch `patched-master`.
This branch is synced with the upstream master and it has only 3 of our
commits applied to it. The goal of the `patched-master` branch is to
merge it in upstream Noir `master` branch.

With this update we had to do several changes to our code.

In a previous commit we removed support for the old fv annotations
syntax. In this commit we are fully removing the dead code associated
with it.

Noir has changed how bitshits work. They no longer require the rhs to be
of type `u8`. They now require the lhs and rhs to be of the same type.
We addressed those changes in our type checking logic for the fv
annotations.

With this Noir dependency version update, we no longer support the
implication operator inside of functions' bodies. We still support it
inside of fv annotations.