feat(fv): Verify files with no main()
Before you couldn't formal verify files which didn't have a main
function. This was because our compilation pipeline assumed that a file
must always have a main function as an entry point. This is the default
assumption in the Noir compiler also.

Now you can verify all functions from a file by providing the file path
as an argument to `verno`. The --help manual of `verno` was also
properly updated. This was made possible by changing our compilation
pipeline to be able to take an function id as an entry point of the
program.

Technical debt: You can not verify generic functions from files which
have no main function. This is because our current workflow expects all
functions to have been monomorphized. When you pass a generic function to
the pipeline as an entry point of the program, there is no way for that
generic function to have been monomorphized.

Added tests which can be only verified using the new CLI command
feature.

Example command:
cargo run -- fv --program-dir
      test_programs/formal_verify_file_mode/cross_file/
      test_programs/formal_verify_file_mode/cross_file/src/helpers.nr